# Fix eviction performance for the scenario with disable automatic prefix caching